License project under Mozilla Public License 2.0
It is important to define the license under which a software project is released in order to permit its use.

Previously, this project did not contain any information about licensing. It is now offered under the formal open source
Mozilla Public License 2.0.

We did our best to assess the intellectual property of this codebase to ensure that all contributors granted copyright
to Arduino SA. For any inquiry, report, or claim please contact us at license@arduino.cc.

The 3rd party CapacitiveSensor library is vendored in the project. That library is (and remains) under the permissive
MIT license, as is already documented in its files.